Encapsulate MAAT-API-specific court application logic
We namespace this CourtApplication class with MaatApi to signify
that its responsibility is to format a payload for Maat Api, as
opposed to the HmctsCommonPlatform::CourtApplication class that
acts solely as a wrapper around court application data coming
from Common Platform.
